// src/Interfaces/Order.tsx
// Generated by https://quicktype.io

export interface OrderDetails {
  idOrden: number;
  idProductoOrden: number;
  Producto: string;
  descripcion: string;
  cantidad: number;
  precioUnitario: number;
  precioTotal: number;
}

export interface Order {
  idOrden: number;
  fecha: string;
  cliente: string;
  total: number | null;
}
// Generated by https://quicktype.io

export interface ClientOrder {
  idCliente: number;
  nombre: string;
  direccion: string;
  telefono: string;
  fecha: string;
  fechaEntrega: null;
  total: number;
}

// Generated by https://quicktype.io

export interface ProductoOrden {
  idProductoOrden: number;
  idOrden: number;
  idProductoMedida: number;
  cantidad: number;
  precioUnitario: number;
  precioTotal: number;
}

// Generated by https://quicktype.io

export interface AddProductoOrden {
  idOrden: number;
  idProductoMedida: number;
  cantidad: number;
  precioUnitario: number;
}